requirements.txt
# Protobuf (compatible with OpenTelemetry and google-genai)
protobuf>=5.26.1,<6.0

# Google Generative AI (modern SDK)
google-genai==1.7.0
langchain-google-genai==2.1.10

# OpenTelemetry stack
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-proto==1.31.0
opentelemetry-util-http==0.52b0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0

# Other Google libraries
google-api-core>=2.19.0,<3.0.0
google-cloud-storage>=2.15.0,<3.0.0

# CrewAI (updated versions for compatibility)
crewai>=0.165.1
crewai-tools>=0.65.0

# FastAPI stack
fastapi==0.110.3
Jinja2==3.1.4
jsonschema==4.22.0
uvicorn==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.20

# LangChain ecosystem (upgraded for compatibility with google-genai)
langchain-core>=0.3.75,<0.4.0
langchain-community>=0.3.5,<0.4.0
langsmith>=0.1.67

# Other utilities
oauthlib==3.2.2
openai>=1.68.2
pydantic==2.7.4
pypdf>=5.0.0,<6.0.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0

